requirements.txt
streamlit==1.38.0
faiss-cpu==1.8.0.post1
pandas==2.2.3
numpy==1.26.4
sentence-transformers==3.1.1
